requirements.txt
pip
Flask
cassandra-driver
requests
requests - cache
